Moving code out of sample app
