+++ stream-chat-android-ui-components-sample/src/main/kotlin/io/getstream/chat/ui/sample/application/ChatInitializer.kt
@@ -26,6 +26,7 @@ import io.getstream.chat.android.client.notifications.handler.NotificationHandle
 import io.getstream.chat.android.markdown.MarkdownTextTransformer
 import io.getstream.chat.android.offline.plugin.configuration.Config
 import io.getstream.chat.android.offline.plugin.factory.StreamOfflinePluginFactory
+import io.getstream.chat.android.offline.repository.realm.initialization.configureRealm
 import io.getstream.chat.android.pushprovider.firebase.FirebasePushDeviceGenerator
 import io.getstream.chat.android.pushprovider.huawei.HuaweiPushDeviceGenerator
 import io.getstream.chat.android.pushprovider.xiaomi.XiaomiPushDeviceGenerator
@@ -34,6 +35,7 @@ import io.getstream.chat.android.state.plugin.factory.StreamStatePluginFactory
 import io.getstream.chat.android.ui.ChatUI
 import io.getstream.chat.ui.sample.BuildConfig
 import io.getstream.chat.ui.sample.feature.HostActivity
+import io.getstream.chat.android.offline.repository.realm.repository.factory.RealmRepositoryFactory
 
 class ChatInitializer(private val context: Context) {
 
@@ -69,7 +71,9 @@ class ChatInitializer(private val context: Context) {
         val offlinePluginFactory = StreamOfflinePluginFactory(
             Config(userPresence = true, persistenceEnabled = true),
             context
-        )
+        ).apply {
+            setCustomRepositoryFactory(RealmRepositoryFactory(configureRealm()))
+        }
 
         val statePluginFactory = StreamStatePluginFactory(
             config = StatePluginConfig(
